OSIS-1255 Save the internship subtype
